backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
httpx==0.28.1
azure-identity==1.19.0
azure-storage-blob==12.24.1
aiohttp>=3.9.0
pydantic-settings==2.7.1

# Glossary-enhanced translation (LLM + text extraction)
openai>=1.40.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
openpyxl>=3.1.0
python-pptx>=1.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
respx==0.22.0
